pyproject.toml
[tool.poetry]
name = "Organization Manager"
version = "0.1.0"
description = ""
authors = [""]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.2"
pillow = "^10.2.0"
django-environ = "^0.11.2"
crispy-bootstrap5 = "^2023.10"
django-crispy-forms = "^2.1"
django-extensions = "^3.2.3"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
pdfkit = "^1.0.0"
wkhtmltopdf = "^0.2"
xhtml2pdf = "^0.2.15"
django-taggit = "^5.0.1"
reportlab = "4.0.9"
drf-yasg = "^1.21.7"
django-storages = "^1.14.3"
boto3 = "^1.34.140"



[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"
django-debug-toolbar = "^4.3.0"
pre-commit = "^3.6.2"
factory-boy = "^3.3.0"
coverage = "^7.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
